Moved global types to their own file and adjusted all files accordingly
