+++ src/sceneManager.ts
@@ -1,47 +1,24 @@
 import type CustomScene from "./customScene";
-import type { theme } from "./stateManager";
-
-export type scene = {
-  name: string;
-  sceneClass: typeof CustomScene;
-};
-
-type sceneManagerProps = {
-  scenes: scene[];
-  index: number;
-  numberOfFrequencies: number;
-  themes: theme[];
-  currentThemeIndex: number;
-};
+import type { Theme, Scene, SceneManagerProps } from "./types";
 
 export default class SceneManager {
-  private _scenes: scene[] = [];
-  private themes: theme[] = [];
+  private _scenes: Scene[] = [];
+  private themes: Theme[] = [];
   private currentThemeIndex: number;
   public currentScene: CustomScene;
   public currentSceneIndex: number = 0;
   private numberOfFrequencies: number;
 
-  constructor({
-    scenes,
-    index,
-    numberOfFrequencies,
-    themes,
-    currentThemeIndex,
-  }: sceneManagerProps) {
+  constructor({ scenes, index, numberOfFrequencies, themes, currentThemeIndex }: SceneManagerProps) {
     this._scenes = scenes;
     this.currentSceneIndex = index;
     this.numberOfFrequencies = numberOfFrequencies;
     this.themes = themes;
     this.currentThemeIndex = currentThemeIndex;
-    this.currentScene = new this._scenes[this.currentSceneIndex].sceneClass(
-      this.numberOfFrequencies,
-      this.themes,
-      this.currentThemeIndex
-    );
+    this.currentScene = new this._scenes[this.currentSceneIndex].sceneClass(this.numberOfFrequencies, this.themes, this.currentThemeIndex);
   }
 
-  get scenes(): scene[] {
+  get scenes(): Scene[] {
     return this._scenes;
   }
 
@@ -52,11 +29,7 @@ export default class SceneManager {
   setCurrentScene(index: number) {
     this.currentScene.destroy();
     this.currentSceneIndex = index;
-    this.currentScene = new this._scenes[index].sceneClass(
-      this.numberOfFrequencies,
-      this.themes,
-      this.currentThemeIndex
-    );
+    this.currentScene = new this._scenes[index].sceneClass(this.numberOfFrequencies, this.themes, this.currentThemeIndex);
   }
 
   setNumberOfFrequencies(n: number) {
