+++ src/scenes/chaoticScene.ts
@@ -1,6 +1,6 @@
 import * as T from "three";
 import CustomScene from "../customScene";
-import type { theme } from "../stateManager";
+import type { Theme } from "../types";
 import { disposeObject } from "../utils/utils";
 
 export default class ChaoticScene extends CustomScene {
@@ -8,7 +8,7 @@ export default class ChaoticScene extends CustomScene {
   private numberOfGroups: number;
   private maxScalar: number;
 
-  constructor(numberOfFrequencies: number, themes: theme[], currentThemeIndex: number) {
+  constructor(numberOfFrequencies: number, themes: Theme[], currentThemeIndex: number) {
     super(numberOfFrequencies, themes, currentThemeIndex);
     this.numberOfGroups = 2;
     this.maxScalar = 0;
